refactor: 의존성을 Hilt로 관리하도록 전환
