+++ app/src/main/java/com/rsupport/mobile1/test/activity/App.kt
@@ -1,17 +1,7 @@
 package com.rsupport.mobile1.test.activity
 
 import android.app.Application
-import coil.ImageLoader
-import com.rsupport.mobile1.test.activity.ui.util.CoilImageLoader
+import dagger.hilt.android.HiltAndroidApp
 
-class App : Application(){
-
-    override fun onCreate() {
-        super.onCreate()
-        imageLoader = CoilImageLoader().getImageLoader(this)
-    }
-
-    companion object {
-        lateinit var imageLoader: ImageLoader
-    }
-}
+@HiltAndroidApp
+class App : Application()
