+++ app/src/main/java/com/rsupport/mobile1/test/activity/data/repository/MainRepository.kt
@@ -4,8 +4,9 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import com.rsupport.mobile1.test.activity.data.model.ImageListPagingSource
 import com.rsupport.mobile1.test.activity.data.network.WebScrapper
+import javax.inject.Inject
 
-class MainRepository(private val webScrapper: WebScrapper) {
+class MainRepository @Inject constructor(private val webScrapper: WebScrapper) {
 
     fun getImageList(pageSize: Int) = Pager(
         PagingConfig(
