+++ build.gradle
@@ -14,6 +14,10 @@ buildscript {
     }
 }
 
+plugins {
+    id 'com.google.dagger.hilt.android' version '2.51.1' apply false
+}
+
 allprojects {
     repositories {
         google()
